% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{.nflverse_download_assets}
\alias{.nflverse_download_assets}
\title{Download nflverse archive assets with gh cli}
\usage{
.nflverse_download_assets(
  release_tag,
  file_type = ".rds",
  download_dir = tempdir(),
  repo = "nflverse/nflverse-data"
)
}
\description{
Download nflverse archive assets with gh cli
}
