% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_cli.R
\name{gh_cli_available}
\alias{gh_cli_available}
\title{This functions tries the gh command in a terminal. If it errors, the gh cli
isn't available on the machine or at least not on the PATH variable}
\usage{
gh_cli_available()
}
\description{
This functions tries the gh command in a terminal. If it errors, the gh cli
isn't available on the machine or at least not on the PATH variable
}
