% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{nflverse_archive}
\alias{nflverse_archive}
\title{Archive nflverse}
\usage{
nflverse_archive(release_name, archive_tag)
}
\arguments{
\item{release_name}{Name of the release}

\item{archive_tag}{The release tag to which data from \code{release_name} should
be uploaded}
}
\description{
Archive nflverse
}
