% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{nflverse_save}
\alias{nflverse_save}
\title{Save files to nflverse release}
\usage{
nflverse_save(
  data_frame,
  file_name,
  nflverse_type,
  release_tag,
  file_types = c("rds", "csv", "parquet", "qs", "csv.gz"),
  repo = "nflverse/nflverse-data"
)
}
\arguments{
\item{data_frame}{data_frame to save}

\item{file_name}{file_name to upload as, without the file extension}

\item{nflverse_type}{metadata: name/information to add to data}

\item{release_tag}{name of release to upload to}

\item{file_types}{one or more of \verb{"rds", "csv", "parquet", "qs", "csv.gz"}}

\item{repo}{repository to upload to, default: \code{"nflverse/nflverse-data"}}
}
\description{
This functions attaches nflverse attributes like type and timestamp, saves
data to a temporary directory in all four of csv, rds, parquet, and qs formats,
and then uploads to nflverse-data repository for a specified release tag.
}
