% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{nflverse_upload}
\alias{nflverse_upload}
\title{Upload to nflverse release}
\usage{
nflverse_upload(...)
}
\arguments{
\item{...}{other args passed to \code{piggyback::pb_upload()}}
}
\description{
Upload to nflverse release
}
