% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{nflverse_upload}
\alias{nflverse_upload}
\title{Upload to nflverse release}
\usage{
nflverse_upload(files, tag)
}
\arguments{
\item{files}{vector of filepaths to upload}

\item{tag}{release name}
}
\description{
Upload to nflverse release
}
