% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{nflverse_upload}
\alias{nflverse_upload}
\title{Upload to nflverse release}
\usage{
nflverse_upload(...)
}
\arguments{
\item{...}{other args passed to \code{piggyback::pb_upload()}}

\item{files}{vector of filepaths to upload}

\item{tag}{release name}

\item{repo}{repository to upload to, default: \code{"nflverse/nflverse-data"}}
}
\description{
Upload to nflverse release
}
